Conflicts:
	automation/source/inc/cmdbasestream.hxx
	automation/source/server/cmdbasestream.cxx
	automation/source/server/retstrm.hxx
	automation/source/testtool/cmdstrm.cxx
	automation/source/testtool/cmdstrm.hxx
	automation/source/testtool/tcommuni.cxx
	basctl/prj/d.lst
	basctl/uiconfig/basicide/toolbar/findbar.xml
	cui/source/dialogs/about.cxx
	cui/source/dialogs/about.src
	cui/source/inc/about.hxx
	extensions/source/abpilot/abpservices.cxx
	extensions/source/dbpilots/dbpservices.cxx
	extensions/source/propctrlr/pcrservices.cxx
	extensions/source/svg/makefile.mk
	forms/Library_frm.mk
	lingucomponent/source/hyphenator/altlinuxhyph/hyphen/hyphenimp.cxx
	lingucomponent/source/spellcheck/spell/sspellimp.cxx
	package/prj/d.lst
	package/source/zipapi/XMemoryStream.cxx
	package/source/zipapi/XMemoryStream.hxx
	setup_native/prj/d.lst
	setup_native/source/win32/customactions/relnotes/makefile.mk
	tools/test/export.map
	wizards/com/sun/star/wizards/common/ConfigGroup.py
	wizards/com/sun/star/wizards/common/ConfigNode.py
	wizards/com/sun/star/wizards/common/Configuration.py
	wizards/com/sun/star/wizards/common/Desktop.py
	wizards/com/sun/star/wizards/common/FileAccess.py
	wizards/com/sun/star/wizards/common/Helper.py
	wizards/com/sun/star/wizards/common/SystemDialog.py
	wizards/com/sun/star/wizards/document/OfficeDocument.py
	wizards/com/sun/star/wizards/fax/FaxDocument.py
	wizards/com/sun/star/wizards/fax/FaxWizardDialog.py
	wizards/com/sun/star/wizards/fax/FaxWizardDialogConst.py
	wizards/com/sun/star/wizards/fax/FaxWizardDialogImpl.py
	wizards/com/sun/star/wizards/fax/FaxWizardDialogResources.py
	wizards/com/sun/star/wizards/letter/LetterDocument.py
	wizards/com/sun/star/wizards/letter/LetterWizardDialog.py
	wizards/com/sun/star/wizards/letter/LetterWizardDialogConst.py
	wizards/com/sun/star/wizards/letter/LetterWizardDialogImpl.py
	wizards/com/sun/star/wizards/letter/LetterWizardDialogResources.py
	wizards/com/sun/star/wizards/text/TextDocument.py
	wizards/com/sun/star/wizards/text/TextFieldHandler.py
	wizards/com/sun/star/wizards/text/TextSectionHandler.py
	wizards/com/sun/star/wizards/text/ViewHandler.py
	wizards/com/sun/star/wizards/ui/UnoDialog.py
	wizards/com/sun/star/wizards/ui/UnoDialog2.py
	wizards/com/sun/star/wizards/ui/WizardDialog.py
	wizards/com/sun/star/wizards/ui/event/CommonListener.py
	wizards/com/sun/star/wizards/ui/event/DataAware.py
	wizards/com/sun/star/wizards/ui/event/RadioDataAware.py
	wizards/com/sun/star/wizards/ui/event/UnoDataAware.py
	wizards/util/helpids.h
	wizards/util/hidother.src
	xmlsecurity/prj/build.lst
	xmlsecurity/prj/d.lst
	xmlsecurity/qa/certext/SanCertExt.cxx

+++ automation/source/testtool/tcommuni.cxx
@@ -170,7 +170,7 @@ sal_uLong GetTTPortConfig()
 
     GETSET( abPortToTalk, "TTPort",
         rtl::OString::valueOf(static_cast<sal_Int32>(TESTTOOL_DEFAULT_PORT)) );
-    return (sal_uLong)abPortToTalk.ToInt64();
+    return (sal_uLong)abPortToTalk.ToInt32();
 }
 
 
@@ -197,7 +197,7 @@ sal_uLong GetUnoPortConfig()
 
     GETSET( abPortToTalk, "UnoPort",
         rtl::OString::valueOf(static_cast<sal_Int32>(UNO_DEFAULT_PORT)) );
-    return (sal_uLong)abPortToTalk.ToInt64();
+    return (sal_uLong)abPortToTalk.ToInt32();
 }
 
 /* vim:set shiftwidth=4 softtabstop=4 expandtab: */
